requirements.txt
torch>=2.0
torchvision>=0.15
transformers>=4.40
pillow>=10.0
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
numpy>=1.24
